Construct elen.dic from "plain wordlists"
elen.dic has 3 components: The greek wordlist, the greek wordlist in
capitals without tonos and the english wordlist.

The problem is that we want to update elen.dic more easily when one
of the greek / english wordlists changes. makedict solves this problem.

In particular, there is now a script (capitals_no_accent) that creates
the wordlist of greek capitals automatically and a Makefile that
assembles elen.dic from the 3 wordlists.

Then elen.dic can be moved to the appropriate place (this is done
by "make install")