Add initial lightproof-based sentence checking
Initial implementation of sentence-level checking for Greek in LibreOffice
using LightProof. LightProof is bundled with LibreOffice starting from version
3.5 and is very correctly putting its emphasis in minimizing false positives
instead of maximizing coverage.

For the beginning I have only implemented 3 simple but useful rules:

    1. english->greek quotes: "ειδικός" -> «ειδικός»
    2. Unmatched close parenthesis:  μια παρένθεση που ξέχασε να ανοίξει)
    3. duplicated words: και και που λες -> και που λες

Of course there is still a lot of work to be done but with a little help this
checker can grow into something very useful, especially when we start to combine
it with hunspell (a version of hunspell that will know a fairly good amount of
stuff about Greek Morphology).