Create the "autocorr" directory during "install"
The produced acor_el-GR.dat file is supposed to go to libregreek/autocorr
but since there is no other file there, git won't add this directory to
the repository.

So, create it dynamically during "installation" since I don't like
using "dummy files" just to add a dir to a repo.